requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Time series analysis
statsmodels>=0.14.0
pmdarima>=2.0.0
prophet>=1.1.0
tslearn>=0.6.0
darts>=0.24.0
sktime>=0.21.0

# Machine learning
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Data processing and utilities
scipy>=1.11.0
joblib>=1.3.0
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
loguru>=0.7.0

# Web interface
streamlit>=1.25.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
